+++ pkg/innerring/processors/settlement/common/details.go
@@ -5,14 +5,9 @@ import (
 )
 
 var (
-	auditPrefix                   = []byte{0x40}
 	basicIncomeCollectionPrefix   = []byte{0x41}
 	basicIncomeDistributionPrefix = []byte{0x42}
 )
-
-func AuditSettlementDetails(epoch uint64) []byte {
-	return details(auditPrefix, epoch)
-}
 
 func BasicIncomeCollectionDetails(epoch uint64) []byte {
 	return details(basicIncomeCollectionPrefix, epoch)